# openapi.yaml
openapi: 3.1.0
info:
  title: CAB432 Video Transcoding API
  version: 1.0.0

paths:
  /auth/login:
    post:
      summary: Log in and receive JWT
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                password:
                  type: string
      responses:
        '200':
          description: JWT returned

  /jobs/transcode:
    post:
      summary: Start a video transcoding job
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                filename:
                  type: string
      responses:
        '202':
          description: Job accepted

  /jobs:
    get:
      summary: List jobs owned by current user
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: pageSize
          in: query
          schema:
            type: integer
      responses:
        '200':
          description: List of jobs

  /jobs/{id}:
    get:
      summary: View a specific job
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Job details

  /jobs/{id}/report:
    get:
      summary: View report for a job (HTML or JSON format)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: format
          in: query
          required: false
          schema:
            type: string
            enum: [json]
      responses:
        '200':
          description: Job report view

  /video/upload:
    post:
      summary: Upload a video file
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                video:
                  type: string
                  format: binary
      responses:
        '200':
          description: File uploaded

  /video/files:
    get:
      summary: List uploaded video files
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of files

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT